requirements.txt
Flask==3.0.0
prometheus_client==0.19.0
elasticsearch==8.10.0
apscheduler==3.10.1
pytest==7.4.0
requests==2.31.0
python-dotenv==1.0.0

